Initial commit with all of Rocket's n_audio and libgultra (minus pfsisplug)

+++ src/os/createthread.c
@@ -11,8 +11,8 @@ void osCreateThread(OSThread *t, OSId id, void (*entry)(void *), void *arg, void
     t->next = NULL;
     t->queue = NULL;
     t->context.pc = (u32)entry;
-    t->context.a0 = (u64)arg;
-    t->context.sp = (u64)sp - 16;
+    t->context.a0 = (s64)(s32)arg; // Double cast gets rid of compiler warning
+    t->context.sp = (s64)(s32)sp - 16;
     t->context.ra = (u64)__osCleanupThread;
     mask = OS_IM_ALL;
     t->context.sr = SR_IMASK | SR_EXL | SR_IE;
